Create class MxRoomEventListener to reduce size of class Room
